qrcodebuilder/requirements.txt
qrcode==7.4.2
Pillow==10.1.0
pypng==0.20220715.0
